Created SeriesBuilder class
